Feature/vault (#53)
## Fixed [WA-424](https://yt.worldr.com/issue/WA-424)

***This is a breaking change!***

This is about getting the new app server to talk to Vault and the PG-TDE via k3s -- local set up.

The server now uses Vault (via k8s auth) to get the key to unseal the TDE database. If Vault is sealed, the app server uses the new vault automatic unsealing serverice.

## Details of commits:

* WIP: Fix release workflow.

* WIP: Get k8s token.

Note that I removed codespell from pre-commit. It is not flexible enough.

* WIP Vault unseal status.

* WIP get KV secrets from Vault.

* WIP sending key to listener.

* Added unseal service API call.